+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Updated
 1. Fixed initial round of _staticcheck_ lint errors and added _staticcheck_ to CI build.
+2. Correct auth flow to suggest `authorise` command on fail
 
 
 ## [0.8.3](https://github.com/uhppoted/uhppoted-app-sheets/releases/tag/v0.8.3) - 2022-12-16